requirements.txt
absl-py==0.5.0
astor==0.7.1
backcall==0.1.0
bleach==2.1.4
cloudpickle==0.5.6
cycler==0.10.0
dask==0.19.2
decorator==4.3.0
-e git+https://github.com/marcoancona/DeepExplain.git#egg=deepexplain
defusedxml==0.5.0
entrypoints==0.2.3
gast==0.2.0
grpcio==1.15.0
h5py==2.8.0
html5lib==1.0.1
ipdb==0.11
ipykernel==5.0.0
ipython==7.0.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.12.1
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
Keras==2.2.2
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
llvmlite==0.25.0
Markdown==3.0.1
MarkupSafe==1.0
matplotlib==3.0.0
mistune==0.8.3
mypy==0.650
mypy-extensions==0.4.1
nbconvert==5.4.0
nbformat==4.4.0
networkx==2.2
notebook==5.7.0
numba==0.40.0
numpy==1.15.4
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.2.0
prometheus-client==0.3.1
prompt-toolkit==2.0.5
protobuf==3.6.1
ptyprocess==0.6.0
Pygments==2.2.0
pyparsing==2.2.1
python-dateutil==2.7.3
pytz==2018.5
PyWavelets==1.0.1
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.1
scikit-image==0.14.0
scikit-learn==0.20.0
scipy==1.1.0
seaborn==0.9.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
tensorboard==1.12.0
tensorflow-gpu==1.12.0
tensorflow-probability==0.5.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.1
toolz==0.9.0
tornado==5.1.1
tqdm==4.26.0
traitlets==4.3.2
typed-ast==1.1.0
umap-learn==0.3.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
